chore: refactor logic out of UI
