+++ src/styles.ts
@@ -1,0 +1,8 @@
+import { css, CSSResultGroup } from 'lit';
+
+export const globalStyles: CSSResultGroup = css`
+  .electricity-schedule-card td {
+    text-align: center;
+  }
+
+`;
